Move features from localboot into generic diskutils package
